Clean up Day creation functions & merge
Pull in the `onCreateNode` code to the node module.
