+++ src/create/index.js
@@ -1,5 +1,7 @@
 export { default as categories } from './categories';
 export { default as gistpens } from './gistpens';
+export * from './node';
 export { default as pages } from './pages';
 export { default as posts } from './posts';
+export * from './schema';
 export { default as tags } from './tags';
