+++ src/create/node.js
@@ -1,11 +1,17 @@
-const { startOfDay, parse, subDays, isBefore } = require('date-fns');
-const crypto = require('crypto');
+import { startOfDay, parse, subDays, isBefore } from 'date-fns';
+import crypto from 'crypto';
+import axios from 'axios';
 
 // @TODO(mAAdhaTTah) include all and paginate
 const earliest = startOfDay(subDays(new Date(), 7));
 const isIncluded = readAt => isBefore(earliest, readAt);
 
-exports.sourceNodes = async ({ node, actions, createNodeId, reporter }) => {
+export const sourceNodes = async ({
+  node,
+  actions,
+  createNodeId,
+  reporter,
+}) => {
   const { createNode, touchNode } = actions;
   const today = startOfDay(new Date());
   const days = [today];
@@ -39,8 +45,67 @@ exports.sourceNodes = async ({ node, actions, createNodeId, reporter }) => {
   }
 };
 
-exports.onCreateNode = async ({ node, actions, createNodeId, reporter }) => {
-  const { createNode, touchNode } = actions;
+let providersP = null;
+
+const getProviders = async () => {
+  const response = await axios.get('https://oembed.com/providers.json');
+
+  return response.data;
+};
+
+const getProviderEndpointForLinkUrl = (linkUrl, providers) => {
+  let transformedEndpoint = {};
+
+  for (const provider of providers || []) {
+    for (const endpoint of provider.endpoints || []) {
+      for (let schema of endpoint.schemes || []) {
+        schema = schema.replace('*', '.*');
+        const regExp = new RegExp(schema);
+        if (regExp.test(linkUrl)) {
+          transformedEndpoint.url = endpoint.url;
+          transformedEndpoint.params = {
+            url: linkUrl,
+            ...provider.params,
+          };
+        }
+      }
+    }
+  }
+
+  return transformedEndpoint;
+};
+
+const getOembed = async (linkUrl, reporter) => {
+  const providers = await providersP;
+  const endpoint = getProviderEndpointForLinkUrl(linkUrl, providers);
+
+  if (!endpoint.url) {
+    reporter.error(`endpoint not found for url: ${linkUrl}`);
+    return null;
+  }
+
+  try {
+    const response = await axios.get(endpoint.url, {
+      params: {
+        format: 'json',
+        ...endpoint.params,
+      },
+    });
+
+    return response.data;
+  } catch (e) {
+    reporter.error(`Request failed for ${endpoint.url}`, e);
+    return null;
+  }
+};
+
+export const onCreateNode = async ({
+  node,
+  actions,
+  createNodeId,
+  reporter,
+}) => {
+  const { createNode, touchNode, createNodeField } = actions;
 
   try {
     let id, readAt;
@@ -96,11 +161,29 @@ exports.onCreateNode = async ({ node, actions, createNodeId, reporter }) => {
           },
         });
         break;
+      case 'wordpress__POST':
+        if (providersP == null) {
+          providersP = getProviders();
+        }
+
+        const {
+          _format_audio_embed: audioUrl,
+          _format_video_embed: videoUrl,
+        } = node.meta;
+
+        const name = 'oembed';
+        const value = {
+          audio: audioUrl ? await getOembed(audioUrl, reporter) : null,
+          video: videoUrl ? await getOembed(videoUrl, reporter) : null,
+        };
+
+        createNodeField({ node, name, value });
+        return;
+      default:
+        return;
     }
 
-    if (id != null) {
-      await touchNode({ nodeId: id });
-    }
+    await touchNode({ nodeId: id });
   } catch (e) {
     reporter.panicOnBuild(`Failed to build links`, e);
   }
